✨ Add shared deadlines to SiDB gate and circuit design
Use a millisecond timeout with an unlimited default. Share one monotonic deadline across circuit gates, candidate generation, pruning, and supported simulators; throw instead of returning partial results.

+++ test/technology/sidb/simulation/engines/test_exhaustive_ground_state_simulation.cpp
@@ -48,7 +48,7 @@ TEST_CASE("ExGS rejects incomplete simulations after the caller deadline", "[exh
     SECTION("Already expired") {}
     SECTION("Expires while enumerating")
     {
-        for (int64_t i = 0; i < 20; ++i)
+        for (int32_t i = 0; i < 20; ++i)
         {
             lyt.assign_sidb({i, 0, 0}, dot_tag::NORMAL);
         }
